# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/engagement/posts/{id}/like": {
      "post": {
        "description": "Toggle like status for a post. If already liked, it will be unliked.",
        "operationId": "EngagementController_toggleLike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Like toggled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "liked": {
                      "type": "boolean",
                      "example": true
                    },
                    "likesCount": {
                      "type": "number",
                      "example": 15
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Like or unlike a post",
        "tags": [
          "Engagement"
        ]
      }
    },
    "/engagement/posts/{id}/likes": {
      "get": {
        "description": "Retrieve paginated list of users who liked a specific post",
        "operationId": "EngagementController_getPostLikes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get users who liked a post",
        "tags": [
          "Engagement"
        ]
      }
    },
    "/engagement/me/likes": {
      "get": {
        "description": "Retrieve paginated list of posts that the authenticated user has liked",
        "operationId": "EngagementController_getAllMyLikedPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get posts liked by authenticated user",
        "tags": [
          "Engagement"
        ]
      }
    },
    "/engagement/posts/{id}/bookmark": {
      "post": {
        "description": "Toggle bookmark status for a post. If already bookmarked, it will be removed.",
        "operationId": "EngagementController_toggleBookmark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookmark toggled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bookmarked": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Bookmark or unbookmark a post",
        "tags": [
          "Engagement"
        ]
      }
    },
    "/engagement/me/bookmarks": {
      "get": {
        "description": "Retrieve paginated list of posts bookmarked by the authenticated user",
        "operationId": "EngagementController_getMyBookmarkedPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get bookmarked posts",
        "tags": [
          "Engagement"
        ]
      }
    },
    "/engagement/posts/{id}/reply": {
      "post": {
        "description": "Create a reply to a post. Supports text content and optional media upload.",
        "operationId": "EngagementController_createReply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Reply content (max 280 characters)",
                    "example": "Great post!",
                    "maxLength": 280
                  },
                  "parentReplyId": {
                    "type": "string",
                    "description": "Parent reply ID if replying to another reply",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                  },
                  "media": {
                    "type": "string",
                    "format": "binary",
                    "description": "Media file to upload (image, video, gif)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reply created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or media file"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a reply to a post",
        "tags": [
          "Engagement"
        ]
      }
    },
...
  },
  "info": {
    "title": "Echo Social Media API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.echo-app.com",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
...
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "visibility": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "likeCount": {
            "type": "number"
          },
          "replyCount": {
            "type": "number"
          },
          "repostCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "mediaCount": {
            "type": "number"
          },
          "author": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaResponseDto"
            }
          },
          "hasLiked": {
            "type": "boolean"
          },
          "hasBookmarked": {
            "type": "boolean"
          },
          "hasReposted": {
            "type": "boolean"
          },
          "hasReplied": {
            "type": "boolean"
          },
          "repostContent": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "number"
          },
          "authorId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "visibility",
          "createdAt",
          "likeCount",
          "replyCount",
          "repostCount",
          "viewCount",
          "mediaCount",
          "author",
          "media",
          "hasLiked",
          "hasBookmarked",
          "hasReposted",
          "hasReplied",
          "repostContent"
        ]
      },
...
    }
  }
}